refactor: move-only: move `FeeEstimateMode` enum to `util/fees.h`
